formatting and ordering
I found it confusing to read about a sample application and the link takes me to the JavaDoc for a class I don't know where to find the code for. I found the code on github before I read about the code later in the readme. I've added a "code, docs" section so that you can go straight to the relevant pages fro the sample section (since it is supposed to be a quick start sample section). I've also sorted out the indentation of the sample java code, the main issue being the braces didn't line up.